Add Logger class
The logger constructor accepts a stream to write to. The _log function logs the severity level of the log message, the action taken during it and a message.
The _log function is meant to be used through the log macro to automate the logging of the filename and line number.

The StaticString class is a wrapper class for the C-style strings used to contain the message.

The Utility class contains a function to ease the conversion from numbers to strings.