Assert potential new implementation for storing the free blocks
Until this commit the linked lists for the free blocks were stored in a statically allocated std::array<Block*, 32>.
This commit adds a new array of linked lists that is stored in the allocator and has a dynamic length equal to the levels in the tree.
This commit ONLY ASSERTS THE CORRECTNESS of the new implementation by running the two simultaneously.